Split top bar surface content
